requirements.txt
kafka-python
pandas
s3fs
boto3
